added error for number format

+++ src/main/java/me/jarenandrews/botListeners.java
@@ -43,34 +43,45 @@ public class botListeners extends ListenerAdapter {
                     //get message
                     String input = event.getMessage().getContentRaw();
                     String[] messageArr = input.split(",");
+                    //error if array is not size 3
+                    if(messageArr.length != 3) {
+                        event.getChannel().sendMessage("Message structure should be *Message*, *font*, *font size* with the commas between, idiot. " + event.getMessage().getAuthor().getAsMention()).queue();
+                        System.out.println("User " + event.getAuthor().getName() + " used incorrect syntax when trying to use the bot. shame.");
+                        return;
+                    }
                     //split parameters
                     String message = messageArr[0];
                     String font = messageArr[1].strip();
-                    int fontSize = Integer.parseInt(messageArr[2].strip());
-                    System.out.println("Message to write: " + message);
-                    //get image file
-                    Message.Attachment att = event.getMessage().getAttachments().get(0);
-                    //create temp file to draw on and send
-                    File temp = new File("draw.png");
-                    //noinspection deprecation
-                    att.downloadToFile(temp).get(); //don't know current way, oh well
-                    BufferedImage img = ImageIO.read(temp);
-                    Graphics2D g = img.createGraphics();
-                    //caption block - color coming soon
-                    Font meme = new Font(font, Font.PLAIN, fontSize);
-                    g.setFont(meme);
-                    FontRenderContext frc = g.getFontRenderContext();
-                    float messageWidth = (float)meme.getStringBounds(message, frc).getWidth();
-                    g.drawString(message, (img.getWidth()-messageWidth)/2, img.getHeight()-(int)(img.getHeight()/6.0));
-                    //end caption block
-                    //write the changes to the image
-                    ImageIO.write(img, "png", temp);
-                    //send image to the channel
-                    event.getChannel().sendFiles(AttachedFile.fromData(temp)).queue();
-                    att.close(); //not sure that this helps
-                    //attempt deletion
-                    if(!temp.delete()) { //can't always delete? overwrites anyway so no big deal
-                        System.out.println("Unable to delete " + temp.getName());
+                    try {
+                        int fontSize = Integer.parseInt(messageArr[2].strip());
+                        System.out.println("Message to write: " + message);
+                        //get image file
+                        Message.Attachment att = event.getMessage().getAttachments().get(0);
+                        //create temp file to draw on and send
+                        File temp = new File("draw.png");
+                        //noinspection deprecation
+                        att.downloadToFile(temp).get(); //don't know current way, oh well
+                        BufferedImage img = ImageIO.read(temp);
+                        Graphics2D g = img.createGraphics();
+                        //caption block - color coming soon
+                        Font meme = new Font(font, Font.PLAIN, fontSize);
+                        g.setFont(meme);
+                        FontRenderContext frc = g.getFontRenderContext();
+                        float messageWidth = (float)meme.getStringBounds(message, frc).getWidth();
+                        g.drawString(message, (img.getWidth()-messageWidth)/2, img.getHeight()-(int)(img.getHeight()/6.0));
+                        //end caption block
+                        //write the changes to the image
+                        ImageIO.write(img, "png", temp);
+                        //send image to the channel
+                        event.getChannel().sendFiles(AttachedFile.fromData(temp)).queue();
+                        att.close(); //not sure that this helps
+                        //attempt deletion
+                        if(!temp.delete()) { //can't always delete? overwrites anyway so no big deal
+                            System.out.println("Unable to delete " + temp.getName());
+                        }
+                    } catch (NumberFormatException e) {
+                        event.getChannel().sendMessage("Font size was not a number, idiot. " + event.getMessage().getAuthor().getAsMention()).queue();
+                        System.out.println("User " + event.getAuthor().getName() + " did not send a number when trying to use the bot. shame.");
                     }
                 } catch (IOException | ExecutionException | InterruptedException e) {
                     e.printStackTrace();
@@ -78,5 +89,4 @@ public class botListeners extends ListenerAdapter {
             }
         }
     }
-
 }
